Move form optionsto form type. Create macros for empty forms
