+++ src/Shop/AdminBundle/Controller/ProductController.php
@@ -8,9 +8,6 @@ use Shop\AdminBundle\Form\ProductType;
 use Shop\CommonBundle\Controller\CommonController;
 use Shop\WebSiteBundle\Entity\Product;
 use Shop\WebSiteBundle\Service\ProductService;
-use Symfony\Component\Form\FormEvent;
-use Symfony\Component\Form\FormEvents;
-use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -26,98 +23,20 @@ class ProductController extends CommonController
     protected $productService;
 
     /**
-     * Creates a new Product entity.
+     * Displays a form to create a new Product entity.
      *
-     * @Route("/", name="admin_product_create")
-     * @Method("POST")
-     * @Template("AdminBundle:Product:new.html.twig")
+     * @Route("/new", name="admin_product_new")
+     * @Template()
      */
-    public function createAction(Request $request)
+    public function newAction(Request $request)
     {
         $entity = $this->productService->createProduct();
-        $form = $this->createProductForm($entity);
+        $form = $this->createForm(new ProductType(), $entity);
         $form->handleRequest($request);
         if ($form->isValid()) {
             $this->productService->save($entity);
             return $this->redirect($this->generateUrl('admin_product_show', ['id' => $entity->getId()]));
         }
-        return [
-            'entity' => $entity,
-            'form' => $form->createView(),
-        ];
-    }
-
-    /**
-     * @param Product $entity
-     * @return mixed
-     */
-    public function createProductForm(Product $entity)
-    {
-        $entityId = $entity->getId();
-
-        if ($entityId) {
-            $submitLabel = "Update";
-            $submitAction = $this->generateUrl('admin_product_update', ['id' => $entity->getId()]);
-        } else {
-            $submitLabel = "Create";
-            $submitAction = $this->generateUrl('admin_product_create');
-        }
-
-        $form = $this->createFormBuilder(new ProductType(), $entity, [
-            'action' => $submitAction,
-            'method' => 'POST'
-        ])->addEventListener(FormEvents::SUBMIT, function ($event) {
-            $this->onFormSubmit($event);
-        }, 900)
-            ->getForm()
-            ->add('submit', 'submit', ['label' => $submitLabel]);
-        return $form;
-    }
-
-    /**
-     * @param $event
-     */
-    protected function onFormSubmit(FormEvent  $event)
-    {
-        /**
-         * @var Product
-         */
-        $product = $event->getData();
-
-        /**
-         * @var UploadedFile
-         */
-        $file = $product->getFile();
-        if ($file instanceof UploadedFile) {
-            $product->setImage($this->moveUploadedFile($file));
-        }
-    }
-
-    /**
-     * @param UploadedFile $file
-     * @return string
-     */
-    protected function moveUploadedFile(UploadedFile $file)
-    {
-        $dirName = sprintf('images/', time());
-        $fileName = sprintf('%d_%s', time(), $file->getClientOriginalName());
-        //$file->move($dirName, $fileName);
-        copy($file->getPathname(), $dirName . $fileName);
-        //unlink($file->getPathname());
-        return $dirName . $fileName;
-    }
-
-    /**
-     * Displays a form to create a new Product entity.
-     *
-     * @Route("/new", name="admin_product_new")
-     * @Method("GET")
-     * @Template()
-     */
-    public function newAction()
-    {
-        $entity = $this->productService->createProduct();
-        $form = $this->createProductForm($entity);
         return [
             'entity' => $entity,
             'form' => $form->createView(),
@@ -160,32 +79,12 @@ class ProductController extends CommonController
      * @Route("/{id}/edit", name="admin_product_edit", requirements={
      *     "id": "\d+"
      * })
-     * @Method("GET")
      * @Template()
      */
-    public function editAction($id)
+    public function editAction(Request $request, $id)
     {
         $entity = $this->getProductById($id);
-        $editForm = $this->createProductForm($entity);
-        return [
-            'entity' => $entity,
-            'edit_form' => $editForm->createView(),
-        ];
-    }
-
-    /**
-     * Edits an existing Product entity.
-     *
-     * @Route("/{id}", name="admin_product_update", requirements={
-     *     "id": "\d+"
-     * })
-     * @Method("POST")
-     * @Template("AdminBundle:Product:edit.html.twig")
-     */
-    public function updateAction(Request $request, $id)
-    {
-        $entity = $this->getProductById($id);
-        $editForm = $this->createProductForm($entity);
+        $editForm = $this->createForm(new ProductType(), $entity);
         $editForm->handleRequest($request);
         if ($editForm->isValid()) {
             $this->productService->save($entity);
